feat(buildkite): Adds BuildKite support
Adds BuildKite support
Adds packaging script
Adds pre-commit hooks
Updates maven plugin versions
Renames integration tests to meet Failsafe plugin conventions
Adds RServe JUnit rule to all tests that connect to a RServe
Adds Docker plugin to Maven to create RServe for testing
Adds Log4J2 support to testing

Resolves #AUR-6027 #AUR-6028 #AUR-6029 #AUR-6030
