+++ src/test/resources/log4j2-test.yml
@@ -1,0 +1,25 @@
+---
+Configuration:
+  status: warn
+  name: YAMLConfigTest
+  properties:
+    property:
+      name: filename
+      value: target/test-yaml.log
+  thresholdFilter:
+    level: debug
+  appenders:
+    Console:
+      name: STDOUT
+      PatternLayout:
+        Pattern: "[%-5level] %d{yyyy-MM-dd HH:mm:ss.SSS} [%t] %c{1} - %msg%n"
+  Loggers:
+    logger:
+      - name: au.edu.uq.aurin.statistics
+        level: warn
+        AppenderRef:
+          ref: STDOUT
+    Root:
+      level: warn
+      AppenderRef:
+        ref: STDOUT
